requirements.txt
streamlit==1.37.1
pandas>=2.1
numpy>=1.26
altair>=5.2
python-dateutil>=2.9
requests>=2.32
